Add option to disconnect from VPN and close RDP window
