+++ README.md
@@ -45,3 +45,5 @@ const rdpServer = 'rdp.server.name';
 #### connectToVpnAndOpenRdp(vpnCredentials, rdpServer)
 #### connectToVpn(server, group, username, password)
 #### openRdpWindow(server)
+#### disconnectFromVpn()
+#### closeRdpWindow()
