+++ index.js
@@ -39,7 +39,7 @@ async function openRdpWindow(server) {
         throw new Error('Need to set all credentials for opening Microsoft RDP!');
     }
 
-    // Open RDP connection
+    // Open RDP window
     return new Promise(resolve => {
         // Needs to promisify `spawn` because `spawnSync` does not return after starting `mstsc`
         const process = spawn('cmd.exe', ['/c', 'start', 'mstsc.exe', `/v:${server}`]);
@@ -52,6 +52,30 @@ async function connectToVpnAndOpenRdp(vpnCredentials, rdpServer) {
     await openRdpWindow(rdpServer);
 }
 
+async function disconnectFromVpn() {
+    try {
+        // TODO: passing redundant values because it requires input. Remove this when/if fixed.
+        await ciscoVpn({server: 'noop', username: 'noop', password: 'noop'}).disconnect();
+    } catch (error) {
+        // Check if the VPN client is not connected
+        const trimmedErrorMessage = error.message.replaceAll('VPN>', '').trim();
+        const isVpnAlreadyDisconnected = trimmedErrorMessage.endsWith('The VPN client is not connected.');
+
+        if (!isVpnAlreadyDisconnected) {
+            throw new Error(error);
+        }
+    }
+}
+
+async function closeRdpWindow() {
+    return new Promise(resolve => {
+        const process = spawn('taskkill', ['/im', 'mstsc.exe']);
+        process.on('exit', () => resolve());
+    });
+}
+
 module.exports.connectToVpn = connectToVpn;
 module.exports.openRdpWindow = openRdpWindow;
 module.exports.connectToVpnAndOpenRdp = connectToVpnAndOpenRdp;
+module.exports.disconnectFromVpn = disconnectFromVpn;
+module.exports.closeRdpWindow = closeRdpWindow;
